First part of slaughter. Tests do not pass.
