+++ azbi/v0/version.go
@@ -1,0 +1,6 @@
+package v0
+
+const (
+	kind    = "azbi"
+	version = "v0.1.4"
+)
