+++ globals/enums.go
@@ -1,0 +1,9 @@
+package globals
+
+type Status string
+
+const (
+	Initialized Status = "initialized"
+	Applied     Status = "applied"
+	Destroyed   Status = "destroyed"
+)
