+++ utils/save/save.go
@@ -4,7 +4,6 @@ import (
 	"io/ioutil"
 
 	awsbi "github.com/epiphany-platform/e-structures/awsbi/v0"
-	azbi "github.com/epiphany-platform/e-structures/azbi/v0"
 	azks "github.com/epiphany-platform/e-structures/azks/v0"
 	hi "github.com/epiphany-platform/e-structures/hi/v0"
 	st "github.com/epiphany-platform/e-structures/state/v0"
@@ -12,18 +11,6 @@ import (
 
 func State(path string, state *st.State) error {
 	bytes, err := state.Marshal()
-	if err != nil {
-		return err
-	}
-	err = ioutil.WriteFile(path, bytes, 0644)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func AzBIConfig(path string, config *azbi.Config) error {
-	bytes, err := config.Marshal()
 	if err != nil {
 		return err
 	}
